+++ CHANGES.rst
@@ -8,6 +8,14 @@ Unreleased
 -   Drop support for Python 3.4. (:issue:`1478`)
 -   Remove code that issued deprecation warnings in version 0.15.
     (:issue:`1477`)
+-   Deprecate most top-level attributes provided by the ``werkzeug``
+    module in favor of direct imports. For example, instead of
+    ``import werkzeug; werkzeug.url_quote``, do
+    ``from werkzeug.urls import url_quote. A deprecation warning will
+    show the correct import to use. ``werkzeug.exceptions`` and
+    ``werkzeug.routing`` should also be imported instead of accessed,
+    but for technical reasons can't show a warning.
+    :issue:`2`, :pr:`1640`
 -   Added ``utils.invalidate_cached_property()`` to invalidate cached
     properties. (:pr:`1474`)
 -   Directive keys for the ``Set-Cookie`` response header are not